2/7/2023_3h

+++ app/src/main/java/com/example/doan/CategoryAdapter.java
@@ -60,7 +60,7 @@ public class CategoryAdapter extends RecyclerView.Adapter<CategoryAdapter.MyView
                 public void onClick(View v) {
                     // Truyền dữ liệu về truyện qua intent
                     Intent intent = new Intent(v.getContext(), NovelDetailsActivity.class);
-                    intent.putExtra("story_id",story.getMaT() ); // ví dụ truyền ID của truyện
+                    intent.putExtra("story_id",(int)story.getMaT() ); // ví dụ truyền ID của truyện
                     v.getContext().startActivity(intent);
                 }
             });
